+++ src/components/port/port.js
@@ -3,10 +3,11 @@ import portConfig from './port.config';
 import pubSub from '../../containers/pubSub';
 import board from '../board/board';
 
-export default (game, boards) => {
+export default (player, game) => {
   const port = {
+    // Rename to portController or shipsController?
+    player,
     game,
-    boards,
     init() {
       this.dragStartHandler = this.dragStartHandler.bind(this);
       this.dragEndHandler = this.dragEndHandler.bind(this);
@@ -14,29 +15,40 @@ export default (game, boards) => {
       this.dropHandler = this.dropHandler.bind(this);
       this.rotateHandler = this.rotateHandler.bind(this);
       this.dragStartHandler = this.dragStartHandler.bind(this);
-      this.reset = this.reset.bind(this);
+      this.resetHandler = this.resetHandler.bind(this);
+      this.readyHandler = this.readyHandler.bind(this);
+
+      this.playerBoard =
+        player === 'player_one' ? this.game.playerOneBoard : this.game.playerTwoBoard;
+      this.dropSubscriber = `drop${player.substring(player.indexOf('_'))}`;
+      this.rotateSubscriber = `rotate${player.substring(player.indexOf('_'))}`;
     },
     cacheDOM(element) {
+      this.port = element;
+      this.ports = element.querySelectorAll('.port_ship');
       this.ships = element.querySelectorAll('.ship_box');
       this.resetBtn = element.querySelector('.reset_btn');
+      this.readyBtn = element.querySelector('.ready_btn');
     },
     bindEvents() {
       this.ships.forEach((ship) => {
+        // https://stackoverflow.com/questions/40464690/want-to-perform-different-task-on-mousedown-and-click-event
         ship.addEventListener('mousedown', this.dragStartHandler);
       });
 
-      this.resetBtn.addEventListener('click', this.reset);
+      this.resetBtn.addEventListener('click', this.resetHandler);
+      this.readyBtn.addEventListener('click', this.readyHandler);
 
-      pubSub.subscribe('drop', this.dropHandler);
-      pubSub.subscribe('rotate', this.rotateHandler);
+      pubSub.subscribe(this.dropSubscriber, this.dropHandler);
+      pubSub.subscribe(this.rotateSubscriber, this.rotateHandler);
     },
     render() {
-      const playerOnePort = createElement(portConfig.element);
-      playerOnePort.setAttributes(portConfig.attributes);
-      playerOnePort.setChildren(portConfig.children);
-      this.cacheDOM(playerOnePort);
+      const playerPort = createElement(portConfig.element);
+      playerPort.setAttributes(portConfig.attributes);
+      playerPort.setChildren(portConfig.children);
+      this.cacheDOM(playerPort);
       this.bindEvents();
-      return playerOnePort;
+      return playerPort;
     },
     dragStartHandler(e) {
       console.log('drag start');
@@ -49,7 +61,6 @@ export default (game, boards) => {
 
       this.dragTimer = setTimeout(() => {
         console.log(`adding mousemove and mouseup events`);
-        console.log(document);
         document.addEventListener('mousemove', this.dragMoveHandler);
         document.addEventListener('mouseup', this.dragEndHandler);
         this.draggable.removeEventListener('click', this.rotateHandler);
@@ -69,10 +80,17 @@ export default (game, boards) => {
       const { left, top, width } = this.draggable.getBoundingClientRect();
       const shipLength = parseInt(this.draggable.dataset.length);
       const offSet = (width / shipLength) * 0.5;
+
       const cell = document
         .elementsFromPoint(left + offSet, top + offSet)
         .find((element) => element.classList.contains('cell'));
-      if (cell) {
+
+      const board = document
+        .elementsFromPoint(left + offSet, top + offSet)
+        .find((element) => element.classList.contains('board'));
+      // // console.log(board ? board.parentElement : player);
+      const isPlayerShip = board ? board.parentElement.contains(this.port) : false;
+      if (cell && isPlayerShip) {
         // Dragging over drop zone
         // If draggable is more than 50% over it's 'last' cell
         //  Append the draggable to the cell content container
@@ -82,7 +100,17 @@ export default (game, boards) => {
 
         const id = this.draggable.dataset.id;
         const orientation = this.draggable.dataset.orientation !== 'h';
-        this.game.playerOneBoard.placeShip([x, y], shipLength, orientation, true, false, id);
+        // this.game.playerOneBoard.placeShip([x, y], shipLength, orientation, true, false, id);
+        this.playerBoard.placeShip(
+          [x, y],
+          shipLength,
+          orientation,
+          true,
+          false,
+          id,
+          this.dropSubscriber,
+          this.rotateSubscriber,
+        );
       } else {
         // Dragging over a non drop zone
         if (
@@ -114,7 +142,17 @@ export default (game, boards) => {
         const shipLength = parseInt(this.draggable.dataset.length);
         const id = this.draggable.dataset.id;
         const orientation = this.draggable.dataset.orientation !== 'h';
-        this.game.playerOneBoard.placeShip([x, y], shipLength, orientation, false, false, id);
+        // this.game.playerOneBoard.placeShip([x, y], shipLength, orientation, false, false, id);
+        this.playerBoard.placeShip(
+          [x, y],
+          shipLength,
+          orientation,
+          false,
+          false,
+          id,
+          this.dropSubscriber,
+          this.rotateSubscriber,
+        );
       }
 
       if (!this.dragStart.classList.contains('port_ship') && this.draggable) {
@@ -133,11 +171,19 @@ export default (game, boards) => {
           this.draggable.classList.add('ship_box_transparent');
         } else if (!isDragging && isValidDrop) {
           // If user has stopped dragging over the drop zone
-          console.log(`dragging stopped over the drop zone`);
+          console.log(`dragging ended over the drop zone`);
           cellContent.appendChild(this.draggable);
           this.dropPlaceholder.remove();
           this.draggable.style.left = `-4%`;
           this.draggable.style.top = `-4%`;
+          if (this.resetBtn.classList.contains('inactive')) {
+            this.resetBtn.classList.remove('inactive');
+          }
+
+          if (this.isPortsEmpty()) {
+            console.log(`ALL SHIPS ARE PLACED ON BOARD`);
+            this.readyBtn.classList.remove('inactive');
+          }
         } else if (isDragging && !isValidDrop) {
           // If user is dragging over an invalid drop
           if (this.dropPlaceholder) {
@@ -148,7 +194,7 @@ export default (game, boards) => {
       } else if (!this.cell && isDragging === false) {
         // If user has stopped dragging outside the drop zone
         // Draggable needs to append back to this.dragStart
-        console.log(`dragging stopped outside the drop zone`);
+        console.log(`dragging ended outside the drop zone`);
       }
     },
     rotateHandler(e) {
@@ -167,7 +213,17 @@ export default (game, boards) => {
           const y = parseInt(this.cell.dataset.y);
           const shipLength = parseInt(this.draggable.dataset.length);
           const id = this.draggable.dataset.id;
-          this.game.playerOneBoard.placeShip([x, y], shipLength, newOrientation, false, true, id);
+          // this.game.playerOneBoard.placeShip([x, y], shipLength, newOrientation, false, true, id);
+          this.playerBoard.placeShip(
+            [x, y],
+            shipLength,
+            newOrientation,
+            false,
+            true,
+            id,
+            this.dropSubscriber,
+            this.rotateSubscriber,
+          );
         }
         e.stopImmediatePropagation();
       } else if (e === true && parseInt(this.draggable.dataset.length) > 1) {
@@ -184,22 +240,26 @@ export default (game, boards) => {
         }, 250);
       }
     },
-    reset(e) {
+    resetHandler(e) {
       // Clears board
-      // this.game.playerOneBoard.clearBoard();
-      // console.log(this.boards.playerOne);
-      // console.log(this.playerOneBoard);
-      const resetBtn = e.currentTarget;
-      const playerBoard = resetBtn.closest(
-        resetBtn.closest('.player_one') ? '.player_one' : '.player_two',
+      const playerBoard = this.resetBtn.closest(
+        this.resetBtn.closest('.player_one') ? '.player_one' : '.player_two',
       ).firstChild;
-      const playerPort = resetBtn.closest('.port');
-      console.log(playerBoard);
-      console.log(playerPort);
 
-      this.game.playerOneBoard.clearBoard();
-      playerPort.replaceWith(this.render());
-      playerBoard.replaceWith(board(this.boards.playerOne));
+      this.playerBoard.clearBoard();
+      this.port.replaceWith(this.render());
+      playerBoard.replaceWith(board(this.playerBoard.board));
+    },
+    isPortsEmpty() {
+      return [...this.ports].every((port) => port.firstChild === null);
+    },
+    readyHandler(e) {
+      console.log(e.currentTarget);
+      const isReady = e.currentTarget.dataset.ready === 'true';
+      e.currentTarget.textContent = isReady ? 'Unready' : 'Ready';
+      e.currentTarget.dataset.ready = !isReady;
+      console.log(this.player);
+      pubSub.publish('playerReady', this.player);
     },
   };